requirements.txt
PyMuPDF==1.24.1
python-docx==1.1.0
torch==2.2.2
transformers==4.40.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pymongo==4.7.1
huggingface_hub==0.23.0
streamlit==1.35.0
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.4
requests==2.31.0
